+++ cmd/install.go
@@ -120,3 +120,16 @@ func install(cmd *cobra.Command, args []string) error {
 
 	return nil
 }
+func installPackage(cmd *cobra.Command, args []string) error {
+	allowUnfree := false
+	if cmd.Flags().Changed("allow-unfree") {
+		allowUnfree = true
+	}
+	err := core.NixInstallPackage(args[0], allowUnfree)
+	if err != nil {
+		return err
+	}
+	cmdr.Success.Println(apx.Trans("nixinstall.success"))
+	return nil
+
+}
